# contract/openapi-panel-v1.json
{
  "components": {
    "schemas": {
...
      "AssignmentListResponse": {
        "description": "Full history for a number — this is the timeline the agent page renders.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AssignmentResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "AssignmentListResponse",
        "type": "object"
      },
      "AssignmentResponse": {
        "properties": {
          "agent_id": {
            "format": "uuid",
            "title": "Agent Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "number_id": {
            "format": "uuid",
            "title": "Number Id",
            "type": "string"
          },
          "valid_from": {
            "format": "date-time",
            "title": "Valid From",
            "type": "string"
          },
          "valid_to": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid To"
          }
        },
        "required": [
          "id",
          "number_id",
          "agent_id",
          "valid_from",
          "valid_to",
          "note",
          "created_at"
        ],
        "title": "AssignmentResponse",
        "type": "object"
      },
...
      "CloseAssignmentRequest": {
        "description": "``PATCH /api/v1/assignments/{id}`` — end a holding period.",
        "properties": {
          "note": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "valid_to": {
            "description": "Must be after valid_from.",
            "format": "date-time",
            "title": "Valid To",
            "type": "string"
          }
        },
        "required": [
          "valid_to"
        ],
        "title": "CloseAssignmentRequest",
        "type": "object"
      },
...
      "CreateAssignmentRequest": {
        "description": "``POST /api/v1/numbers/{id}/assignments`` — hand a line to an agent.",
        "properties": {
          "agent_id": {
            "format": "uuid",
            "title": "Agent Id",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "valid_from": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Defaults to now. The period is [valid_from, valid_to).",
            "title": "Valid From"
          },
          "valid_to": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "NULL means open-ended — they hold it now.",
            "title": "Valid To"
          }
        },
        "required": [
          "agent_id"
        ],
        "title": "CreateAssignmentRequest",
        "type": "object"
      },
...
      "CreateNumberRequest": {
        "description": "``POST /api/v1/numbers``.\n\nThe number is normalised to E.164 before it is stored, so the same line\ntyped three ways is one row — the uniqueness lives on the generated\n``phone_key`` column, not on this string.",
        "properties": {
          "e164": {
            "description": "Any format; normalised here.",
            "maxLength": 32,
            "minLength": 6,
            "title": "E164",
            "type": "string"
          },
          "label": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "operator": {
            "anyOf": [
              {
                "maxLength": 32,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "beeline | ucell | mobiuz | uzmobile | other. A reporting axis (R19).",
            "title": "Operator"
          },
          "sim_owner": {
            "default": "company",
            "description": "R10: which SIM relationships are not Bonvi's to keep.",
            "pattern": "^(company|employee)$",
            "title": "Sim Owner",
            "type": "string"
          }
        },
        "required": [
          "e164"
        ],
        "title": "CreateNumberRequest",
        "type": "object"
      },
...
      "EnrolmentCodeResponse": {
        "description": "The code plus everything the admin has to pass on.",
        "properties": {
          "agent_id": {
            "format": "uuid",
            "title": "Agent Id",
            "type": "string"
          },
          "attempt_count": {
            "title": "Attempt Count",
            "type": "integer"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "number_id": {
            "format": "uuid",
            "title": "Number Id",
            "type": "string"
          },
          "redeemed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Redeemed At"
          },
          "revoked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          }
        },
        "required": [
          "id",
          "code",
          "number_id",
          "agent_id",
          "expires_at",
          "redeemed_at",
          "revoked_at",
          "attempt_count",
          "created_at"
        ],
        "title": "EnrolmentCodeResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NumberListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NumberResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "NumberListResponse",
        "type": "object"
      },
      "NumberResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "e164": {
            "title": "E164",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "operator": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operator"
          },
          "phone_key": {
            "description": "Last 9 digits, generated by the database (N37).",
            "title": "Phone Key",
            "type": "string"
          },
          "sim_owner": {
            "title": "Sim Owner",
            "type": "string"
          }
        },
        "required": [
          "id",
          "e164",
          "phone_key",
          "operator",
          "sim_owner",
          "label",
          "is_active",
          "created_at"
        ],
        "title": "NumberResponse",
        "type": "object"
      },
...
      "UpdateNumberRequest": {
        "description": "``PATCH /api/v1/numbers/{id}``. The number itself is not editable.",
        "properties": {
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "label": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "operator": {
            "anyOf": [
              {
                "maxLength": 32,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operator"
          },
          "sim_owner": {
            "anyOf": [
              {
                "pattern": "^(company|employee)$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sim Owner"
          }
        },
        "title": "UpdateNumberRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Generated from the Pydantic schemas. Prefix: /api/v1",
    "title": "BonviCall panel API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/assignments/{assignment_id}": {
      "patch": {
        "description": "Close a holding period. Calls already attributed keep their agent.",
        "operationId": "close_assignment_api_v1_assignments__assignment_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "assignment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Assignment Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseAssignmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Close Assignment",
        "tags": [
          "Numbers"
        ]
      }
    },
...
    "/api/v1/numbers": {
      "get": {
        "operationId": "list_numbers_api_v1_numbers_get",
        "parameters": [
          {
            "in": "query",
            "name": "is_active",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Active"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NumberListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Numbers",
        "tags": [
          "Numbers"
        ]
      },
      "post": {
        "description": "The number is normalised first, so three formats are one row (N37).",
        "operationId": "create_number_api_v1_numbers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNumberRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NumberResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Number",
        "tags": [
          "Numbers"
        ]
      }
    },
    "/api/v1/numbers/{number_id}": {
      "get": {
        "operationId": "get_number_api_v1_numbers__number_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "number_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Number Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NumberResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Number",
        "tags": [
          "Numbers"
        ]
      },
      "patch": {
        "operationId": "update_number_api_v1_numbers__number_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "number_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Number Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNumberRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NumberResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Number",
        "tags": [
          "Numbers"
        ]
      }
    },
    "/api/v1/numbers/{number_id}/assignments": {
      "get": {
        "description": "Full history — this is where the time-boxed mapping becomes visible.",
        "operationId": "list_assignments_api_v1_numbers__number_id__assignments_get",
        "parameters": [
          {
            "in": "path",
            "name": "number_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Number Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Assignments",
        "tags": [
          "Numbers"
        ]
      },
      "post": {
        "description": "409 ``number_already_assigned``, with the current holder in ``detail``.",
        "operationId": "create_assignment_api_v1_numbers__number_id__assignments_post",
        "parameters": [
          {
            "in": "path",
            "name": "number_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Number Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAssignmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Assignment",
        "tags": [
          "Numbers"
        ]
      }
    },
    "/api/v1/numbers/{number_id}/enrolment-code": {
      "post": {
        "description": "Single-use, 24 h. The agent is frozen at issue.",
        "operationId": "issue_enrolment_code_api_v1_numbers__number_id__enrolment_code_post",
        "parameters": [
          {
            "in": "path",
            "name": "number_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Number Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrolmentCodeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Issue Enrolment Code",
        "tags": [
          "Numbers"
        ]
      }
    },
...
  }
}