# server/src/modules/numbers/schemas.py
"""Registered-number and assignment wire schemas (SPEC §3.3, §4.7)."""

from __future__ import annotations

import uuid
from datetime import datetime

from pydantic import BaseModel, ConfigDict, Field, field_validator

from src.core.phone import phone_key, to_e164


class CreateNumberRequest(BaseModel):
    """``POST /api/v1/numbers``.

    The number is normalised to E.164 before it is stored, so the same line
    typed three ways is one row — the uniqueness lives on the generated
    ``phone_key`` column, not on this string.
    """

    e164: str = Field(min_length=6, max_length=32, description="Any format; normalised here.")
    operator: str | None = Field(
        default=None,
        max_length=32,
        description="beeline | ucell | mobiuz | uzmobile | other. A reporting axis (R19).",
    )
    sim_owner: str = Field(
        default="company",
        pattern="^(company|employee)$",
        description="R10: which SIM relationships are not Bonvi's to keep.",
    )
    label: str | None = Field(default=None, max_length=64)

    @field_validator("e164")
    @classmethod
    def _normalise(cls, value: str) -> str:
        normalised = to_e164(value)
        if normalised is None:
            # Fewer than nine digits is never a phone number (N37). Refusing
            # here is right: a registered number with no key can never match a
            # call, so storing it would create a line that silently captures
            # nothing.
            raise ValueError("a registered number needs at least 9 digits")
        return normalised


class UpdateNumberRequest(BaseModel):
    """``PATCH /api/v1/numbers/{id}``. The number itself is not editable."""

    operator: str | None = Field(default=None, max_length=32)
    sim_owner: str | None = Field(default=None, pattern="^(company|employee)$")
    label: str | None = Field(default=None, max_length=64)
    is_active: bool | None = None


class NumberResponse(BaseModel):
    model_config = ConfigDict(from_attributes=True)

    id: uuid.UUID
    e164: str
    phone_key: str = Field(description="Last 9 digits, generated by the database (N37).")
    operator: str | None
    sim_owner: str
    label: str | None
    is_active: bool
    created_at: datetime


class NumberListResponse(BaseModel):
    items: list[NumberResponse]
    total: int


class CreateAssignmentRequest(BaseModel):
    """``POST /api/v1/numbers/{id}/assignments`` — hand a line to an agent."""

    agent_id: uuid.UUID
    valid_from: datetime | None = Field(
        default=None, description="Defaults to now. The period is [valid_from, valid_to)."
    )
    valid_to: datetime | None = Field(
        default=None, description="NULL means open-ended — they hold it now."
    )
    note: str | None = Field(default=None, max_length=255)


class CloseAssignmentRequest(BaseModel):
    """``PATCH /api/v1/assignments/{id}`` — end a holding period."""

    valid_to: datetime = Field(description="Must be after valid_from.")
    note: str | None = Field(default=None, max_length=255)


class AssignmentResponse(BaseModel):
    model_config = ConfigDict(from_attributes=True)

    id: uuid.UUID
    number_id: uuid.UUID
    agent_id: uuid.UUID
    valid_from: datetime
    valid_to: datetime | None
    note: str | None
    created_at: datetime


class AssignmentListResponse(BaseModel):
    """Full history for a number — this is the timeline the agent page renders."""

    items: list[AssignmentResponse]
    total: int


def key_of(raw: str) -> str | None:
    """Re-exported for routers that need to echo the key in a 409 body."""
    return phone_key(raw)